Make uninstall smarter

+++ src/install.sh
@@ -2,11 +2,17 @@
 
 OS=`uname -s`
 
+SELF=`basename "$0"`
 PGM=hwaet
 PGMD="$PGM"d
 INSTALL_DIR='/usr/local'
 DAEMON_NAME='Hostname Responder Daemon'
 # DAEMON_USER is system specific: OpenBSD has an underscore prepended
+
+msg()
+{
+    echo $SELF: $*
+}
 
 installSystemdService()
 (
@@ -38,7 +44,7 @@ installSystemdService()
 
 if [ `whoami` != 'root' ]
 then
-    echo You must be root for this to work. 1>&2
+    msg You must be root for this to work. 1>&2
     exit 1
 fi
 
@@ -72,7 +78,7 @@ case $OS in
         ;;
         
     FreeBSD | NetBSD)
-        echo Don\'t know how on $OS. Please write me.
+        msg Don\'t know how on $OS. Please write me.
         exit
     ;;
     
@@ -101,7 +107,7 @@ case $OS in
         ;;
 
     Linux)
-        echo Recognized Linux
+        msg Recognized Linux
         DAEMON_USER="$PGMD"
         if [ "$OP" = "install" ]
         then
@@ -131,43 +137,74 @@ case $OS in
             fi
         elif [ -d /etc/init.d ] && type service >/dev/null
         then
-            echo Recognized System V Init
+            msg Recognized System V Init
             if [ "$OP" = "install" ]
             then
-                echo Installing init.d script
+                msg Installing init.d script
                 install -p -o root -g root -m u=rwx,g=rx,o=rx $PGMD.sysVinit \
                     /etc/init.d/$PGMD
                 update-rc.d $PGMD defaults
                 service $PGMD start
             else
-                service $PGMD stop
-                echo Removing init.d script
-                update-rc.d $PGMD defaults-disabled
-                rm /etc/init.d/$PGMD
+                if [ -f /etc/init.d/$PGMD ]
+                then
+                    service $PGMD stop
+                    msg Removing init.d script
+                    update-rc.d $PGMD defaults-disabled
+                    rm /etc/init.d/$PGMD
+                else
+                    msg Service already uninstalled.
+                fi
             fi
             
         else
-            echo Unrecognized Linux init system. Write me.
+            msg Unrecognized Linux init system. Write me.
         fi
         
         if [ "$OP" = "uninstall" ]
         then
-            rm $INSTALL_DIR/bin/$PGM $INSTALL_DIR/sbin/$PGMD 
-            userdel $DAEMON_USER
+            if [ -f "$INSTALL_DIR/bin/$PGM" ]
+            then
+                rm "$INSTALL_DIR/bin/$PGM"
+            else
+                msg Client program already removed.
+            fi
+
+            if [ -f "$INSTALL_DIR/bin/$PGMD" ]
+            then
+                rm "$INSTALL_DIR/bin/$PGMD"
+            else
+                msg Daemon program already removed.
+            fi
+
+            if grep "$DAEMON_USER" /etc/passwd >/dev/null
+            then
+                userdel $DAEMON_USER
+            else
+                msg Daemon user already removed.
+            fi
             # Remove mail spool file but not home directory. Doing this with
             # "userdel --remove" would remove the home directory also, but
             # it is a shared directory in this case. So don't do that.
-            MAIL_DIR=`grep '^MAIL_DIR' /etc/login.defs | cut -f2`
-            rm "$MAIL_DIR"
+            # Awk is needed rather than cut. Cut can use spaces or tabs as
+            # delimiters but not both.
+            MAIL_DIR=`grep '^MAIL_DIR' /etc/login.defs | awk '{print $2}'`
+            MAIL_SPOOL="$MAIL_DIR/$DAEMON_USER"
+            if [ -f "$MAIL_SPOOL" ]
+            then
+                rm "$MAIL_SPOOL"
+            else
+                msg Daemon mail spool file already removed.
+            fi
         fi
         ;;
         
     *Win* | *win*)
         # Scoop could be used here.
-        echo $OS? Stop being a looser. Go to http://getfedora.org.
+        msg $OS? Stop being a looser. Go to http://getfedora.org.
         ;;
 
     *)
-        echo Unknown OS: $OS
+        msg Unknown OS: $OS
         ;;
 esac
